refactor: 최종 우승자 계산 및 출력 함수 분리(Class : Winner)
