+++ src/main/java/racingcar/RacingGame.java
@@ -6,12 +6,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class RacingGame {
-    private static final ArrayList<Car> carList = new ArrayList<>();
     private static final int MIN = 0;
     private static final int MAX = 9;
-    private static final int POINT = 4;
-    private static final ArrayList<String> winnerList = new ArrayList<>();
+    private static final int POINT_NUMBER = 4;
+    private static final ArrayList<Car> carList = new ArrayList<>();
     private static final Player player = new Player();
+    private static final Winner winner = new Winner();
 
     public void startGame() {
         String[] carNamesArray = player.startGame();
@@ -20,8 +20,7 @@ public class RacingGame {
         while (player.remainCoin()) {
             this.playOneGame();
         }
-        this.getWinner();
-        this.printWinner();
+        winner.getWinner(carList);
     }
 
     private void createCarList(String[] carNamesArray) {
@@ -41,29 +40,6 @@ public class RacingGame {
     }
 
     private boolean makeRandomNumber() {
-        return Randoms.pickNumberInRange(MIN, MAX) >= POINT;
-    }
-
-    private void getWinner() {
-        int maxPosition = Integer.MIN_VALUE;
-        for (Car car : carList) {
-            if (car.getPosition() > maxPosition) {
-                winnerList.clear();
-                maxPosition = car.getPosition();
-                winnerList.add(car.getName());
-                continue;
-            }
-            if (car.getPosition() == maxPosition) {
-                winnerList.add(car.getName());
-            }
-        }
-    }
-
-    private void printWinner() {
-        if (winnerList.size() != 1) {
-            System.out.println("최종 우승자 : " + String.join(", ", winnerList));
-            return;
-        }
-        System.out.println("최종 우승자 : " + winnerList.get(0));
+        return Randoms.pickNumberInRange(MIN, MAX) >= POINT_NUMBER;
     }
 }
